requirements.txt
openai
Flask
flask-cors
python-dotenv
requests
boto3
pydantic
requests
tiktoken
sendgrid
pydantic
openai-function-call
celery
gunicorn
Werkzeug
retry
